backend/app/services/mcp_client.py
import asyncio
import json
import subprocess
import logging
from typing import Dict, Any, Optional, List
import yaml
import os

from app.core.config import settings

logger = logging.getLogger(__name__)


class MCPSnowflakeClient:
    """
    Client for MCP (Model Context Protocol) integration with Snowflake
    """
    
    def __init__(self, config_path: Optional[str] = None):
        self.config_path = config_path or settings.MCP_SNOWFLAKE_CONFIG_PATH
        self.config = self._load_config()
        
    def _load_config(self) -> Dict[str, Any]:
        """
        Load MCP configuration from YAML file
        """
        try:
            if os.path.exists(self.config_path):
                with open(self.config_path, 'r') as file:
                    return yaml.safe_load(file)
            else:
                # Return default config if file doesn't exist
                return self._get_default_config()
        except Exception as e:
            logger.error(f"Error loading MCP config: {str(e)}")
            return self._get_default_config()
    
    def _get_default_config(self) -> Dict[str, Any]:
        """
        Default MCP configuration
        """
        return {
            "snowflake": {
                "cortex_analyst": {
                    "enabled": True,
                    "endpoint": "https://your-snowflake-account.snowflakecomputing.com"
                },
                "cortex_search": {
                    "enabled": True,
                    "endpoint": "https://your-snowflake-account.snowflakecomputing.com"
                },
                "cortex_agents": {
                    "enabled": True,
                    "endpoint": "https://your-snowflake-account.snowflakecomputing.com"
                }
            },
            "connection": {
                "account": os.getenv("SNOWFLAKE_ACCOUNT", ""),
                "user": os.getenv("SNOWFLAKE_USER", ""),
                "password": os.getenv("SNOWFLAKE_PASSWORD", ""),
                "warehouse": os.getenv("SNOWFLAKE_WAREHOUSE", "COMPUTE_WH"),
                "database": os.getenv("SNOWFLAKE_DATABASE", ""),
                "schema": os.getenv("SNOWFLAKE_SCHEMA", "PUBLIC")
            }
        }
    
    async def execute_query(self, query_request: str) -> Dict[str, Any]:
        """
        Execute a natural language query using Snowflake Cortex Analyst
        """
        try:
            # Use Cortex Analyst for natural language queries
            result = await self._call_cortex_analyst(query_request)
            return result
        except Exception as e:
            logger.error(f"Error executing MCP query: {str(e)}")
            raise
    
    async def search_data(self, search_query: str, context: Optional[str] = None) -> Dict[str, Any]:
        """
        Search data using Snowflake Cortex Search
        """
        try:
            result = await self._call_cortex_search(search_query, context)
            return result
        except Exception as e:
            logger.error(f"Error in MCP search: {str(e)}")
            raise
    
    async def invoke_agent(self, agent_name: str, input_data: Dict[str, Any]) -> Dict[str, Any]:
        """
        Invoke a specific Snowflake Cortex Agent
        """
        try:
            result = await self._call_cortex_agent(agent_name, input_data)
            return result
        except Exception as e:
            logger.error(f"Error invoking MCP agent: {str(e)}")
            raise
    
    async def _call_cortex_analyst(self, query_request: str) -> Dict[str, Any]:
        """
        Call Snowflake Cortex Analyst via MCP
        """
        # This is a placeholder implementation
        # In a real implementation, you would use the actual MCP client library
        # or make direct API calls to Snowflake Cortex Analyst
        
        try:
            # Simulate MCP call to Cortex Analyst
            # In reality, this would use the snowflake-labs/mcp client
            
            # For now, return a mock response
            return {
                "query": query_request,
                "response": f"Mock Cortex Analyst response for: {query_request}",
                "data": [
                    {"column1": "value1", "column2": "value2"},
                    {"column1": "value3", "column2": "value4"}
                ],
                "metadata": {
                    "execution_time_ms": 150,
                    "rows_returned": 2,
                    "service": "cortex_analyst"
                }
            }
        except Exception as e:
            logger.error(f"Cortex Analyst call failed: {str(e)}")
            raise
    
    async def _call_cortex_search(self, search_query: str, context: Optional[str] = None) -> Dict[str, Any]:
        """
        Call Snowflake Cortex Search via MCP
        """
        try:
            # Simulate MCP call to Cortex Search
            return {
                "query": search_query,
                "context": context,
                "results": [
                    {
                        "content": f"Search result 1 for: {search_query}",
                        "relevance_score": 0.95,
                        "source": "table1"
                    },
                    {
                        "content": f"Search result 2 for: {search_query}",
                        "relevance_score": 0.87,
                        "source": "table2"
                    }
                ],
                "metadata": {
                    "total_results": 2,
                    "search_time_ms": 75,
                    "service": "cortex_search"
                }
            }
        except Exception as e:
            logger.error(f"Cortex Search call failed: {str(e)}")
            raise
    
    async def _call_cortex_agent(self, agent_name: str, input_data: Dict[str, Any]) -> Dict[str, Any]:
        """
        Call Snowflake Cortex Agent via MCP
        """
        try:
            # Simulate MCP call to Cortex Agent
            return {
                "agent": agent_name,
                "input": input_data,
                "output": f"Mock agent {agent_name} response",
                "metadata": {
                    "execution_time_ms": 200,
                    "service": "cortex_agents"
                }
            }
        except Exception as e:
            logger.error(f"Cortex Agent call failed: {str(e)}")
            raise
    
    def get_available_services(self) -> List[str]:
        """
        Get list of available MCP services
        """
        services = []
        snowflake_config = self.config.get("snowflake", {})
        
        for service, config in snowflake_config.items():
            if config.get("enabled", False):
                services.append(service)
        
        return services
    
    def test_connection(self) -> Dict[str, Any]:
        """
        Test MCP connection to Snowflake
        """
        try:
            # This would test the actual connection in a real implementation
            connection_config = self.config.get("connection", {})
            
            # Check if required connection parameters are provided
            required_params = ["account", "user", "password"]
            missing_params = [param for param in required_params if not connection_config.get(param)]
            
            if missing_params:
                return {
                    "status": "failed",
                    "error": f"Missing connection parameters: {', '.join(missing_params)}"
                }
            
            return {
                "status": "success",
                "message": "MCP connection test successful",
                "services": self.get_available_services()
            }
            
        except Exception as e:
            return {
                "status": "failed",
                "error": str(e)
            }