backend/requirements.txt
# FastAPI and dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Async task queue
celery==5.3.4
redis==5.0.1

# AI/LLM
openai==1.3.7
azure-ai-ml==1.11.1
azure-identity==1.15.0

# WebSockets
websockets==12.0

# File processing
pypdf==3.17.1
python-docx==1.1.0

# Utils
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
jinja2==3.1.2
aiofiles==23.2.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0